put everything in norm

+++ ft_printint.c
@@ -6,15 +6,15 @@
 /*   By: mkettab <mkettab@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/18 19:35:21 by mkettab           #+#    #+#             */
-/*   Updated: 2024/12/05 18:25:53 by mkettab          ###   ########.fr       */
+/*   Updated: 2024/12/05 19:26:17 by mkettab          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int ft_putint(long d)
+int	ft_putint(long d)
 {
-	long first;
+	long	first;
 
 	first = d;
 	if (d < 0)
@@ -22,17 +22,17 @@ int ft_putint(long d)
 		d = -d;
 		ft_putchar('-');
 	}
-	if(d >= 10)
+	if (d >= 10)
 	{
 		ft_putint(d / 10);
 		ft_putint(d % 10);
 	}
-	if(d < 10)
+	if (d < 10)
 		ft_putchar(d + '0');
 	return (ft_intlen(first));
 }
 
-int ft_putunsint(long d)
+int	ft_putunsint(long d)
 {
 	if (d >= 10)
 	{
@@ -46,20 +46,20 @@ int ft_putunsint(long d)
 
 int	ft_puthex(unsigned long c, int low)
 {
-	unsigned long long temp;
-	char *base;
+	unsigned long	temp;
+	char			*base;
 
 	temp = c;
-	if(low == 1)
+	if (low == 1)
 		base = "0123456789abcdef";
-	if(low == 0)
+	if (low == 0)
 		base = "0123456789ABCDEF";
 	if (c >= 16)
 	{
 		ft_puthex(c / 16, low);
 		ft_puthex(c % 16, low);
 	}
-	if(c < 16)
+	if (c < 16)
 		ft_putchar(base[c]);
 	return (ft_inthexlen(temp));
 }
